+++ src/index.js
@@ -158,11 +158,12 @@ var $ = function(selector, context){
     return new $.init(selector, context);
 };
 
-$.version = "v2.0.0";
-$.build_time = "21.04.2024, 17:20:12";
-$.info = () => console.info(`%c M4Q %c v${$.version} %c ${$.build_time} `, "color: pink; font-weight: bold; background: #800000", "color: white; background: darkgreen", "color: white; background: #0080fe;")
+$.version = "2.0.0-rc1";
+$.build_time = "21.04.2024, 19:59:41";
+$.info = () => console.info(`%c M4Q %c v${$.version} %c ${$.build_time} `, "color: white; font-weight: bold; background: #fd6a02", "color: white; background: darkgreen", "color: white; background: #0080fe;")
 
 $.fn = $.prototype = {
+    [Symbol.isConcatSpreadable]: true,
     constructor: $,
     length: 0,
     uid: "",
@@ -170,8 +171,16 @@ $.fn = $.prototype = {
     push: [].push,
     sort: [].sort,
     splice: [].splice,
+    slice: [].slice,
     indexOf: [].indexOf,
-    reverse: [].reverse
+    lastIndexOf: [].lastIndexOf,
+    reverse: [].reverse,
+    map: [].map,
+    pop: [].pop,
+    shift: [].shift,
+    unshift: [].unshift,
+    concat: [].concat,
+    includes: [].includes,
 };
 
 $.extend = $.fn.extend = function(){
@@ -2311,7 +2320,14 @@ $.fn.extend({
     },
 
     id: function(val){
-        return this.length ? $(this[0]).attr("id", val) : undefined;
+        if (typeof val === "undefined") {
+            return this.length === 1
+                ? $(this[0]).attr("id")
+                : this.map(el => $(el).attr("id"))
+        }
+        return this.each(function(){
+            $(this).attr("id", val);
+        });
     }
 });