+++ test/attributes.js
@@ -1,0 +1,5 @@
+import {$} from "../dist/m4q.esm.js"
+
+$.info()
+$("#my-div").attr("id", "my-div2")
+console.log($("h1").id())